El administrador depende de los permisos de Escritura para modificar

+++ php/tablaAdmin.php
@@ -53,39 +53,52 @@
                 
                 <?php 
                 
-                    $resultado = mysqli_query($conexion, $datos); 
+                    $resultado = mysqli_query($conexion, $datos);
+
+                    //Comprobar el permiso de Escritura
+                    $escrituraCuenta = "SELECT escritura FROM registros WHERE email = '$usuario'";
+                    $escrituraCheck = mysqli_query($conexion, $escrituraCuenta);
+                    $escritura = mysqli_fetch_assoc($escrituraCheck);
+                    #var_dump($escritura);
                     
                     while($row = mysqli_fetch_assoc($resultado)){
-                ?>
+                    ?>
 
-                <tr>
-                    <!--<td><?php echo $row['id']; ?></td>-->
-                    <td><?php echo $row['nombre']; ?></td>
-                    <td><?php echo $row['email']; ?></td>
-                    <td><?php echo $row['telefono']; ?></td>
-                    <td><?php echo $row['contrasenia']; ?></td>
-                    <td><?php echo $row['rol']; ?></td>
-                    <td><?php echo $row['lectura']; ?></td>
-                    <td><?php echo $row['escritura']; ?></td>
-                    <td><?php echo $row['status']; ?></td>
-                    <td>
-                        <button class="actualizar">
-                            <?php echo "<a href='editarAdmin.php?id=".$row['id']."'>Editar</a>"; ?>
-                            <i class="fa-solid fa-pencil"></i>
-                        </button>
-                    </td>
-                    <td>
-                        <button class="eliminar">
-                            <?php echo "<a href='eliminar.php?id=".$row['id']."' onclick='return confirmar()'>Eliminar</a>"; ?>
-                            <i class="fa-solid fa-eraser"></i>
-                        </button>
-                    </td>
-                </tr>
+                    <tr>
+                        <!--<td><?php echo $row['id']; ?></td>-->
+                        <td><?php echo $row['nombre']; ?></td>
+                        <td><?php echo $row['email']; ?></td>
+                        <td><?php echo $row['telefono']; ?></td>
+                        <td><?php echo $row['contrasenia']; ?></td>
+                        <td><?php echo $row['rol']; ?></td>
+                        <td><?php echo $row['lectura']; ?></td>
+                        <td><?php echo $row['escritura']; ?></td>
+                        <td><?php echo $row['status']; ?></td>
+                        
+                        <?php
+                            if($escritura['escritura'] == "Si"){
+                                ?>
+                                    <td>
+                                        <button class="actualizar">
+                                            <?php echo "<a href='editarAdmin.php?id=".$row['id']."'>Editar</a>"; ?>
+                                            <i class="fa-solid fa-pencil"></i>
+                                        </button>
+                                    </td>
+                                    <td>
+                                        <button class="eliminar">
+                                            <?php echo "<a href='eliminar.php?id=".$row['id']."' onclick='return confirmar()'>Eliminar</a>"; ?>
+                                            <i class="fa-solid fa-eraser"></i>
+                                        </button>
+                                    </td>
+                                <?php
+                            }
+                        ?>
+                    </tr>
                     
-                <?php 
-                    }
-                    mysqli_free_result($resultado);
-                ?>
+                    <?php 
+                        }
+                        mysqli_free_result($resultado);
+                    ?>
                 
             </table>
         </div>
